New Statistics UI, more statistics (#1075)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

This PR adds a new UI to view round end statistics, instead of them
being just parsed into the chat. It also adds a new link next to the
player list and commend links at the round end, so the players can
revisit statistics UI anytime they want to.

![End Round
Stats](https://github.com/user-attachments/assets/7fc0846d-926d-4575-9537-ab556bdf350c)

Lastly, 14 new statistics were added:

- Noble deaths
- Trees cut
- Prayers made
- Fish caught
- Number of pickpockets
- Masterworks forged
- Mammons taxed
- Organs eaten
- Kisses made
- Laughs had
- Graves consecrated
- Potions brewed
- Astrata revivals
- Plants harvested

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

Better way to view statistics, more stats that could even be utilized
later. (for storytellers for example)

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->