Half-elf hair fix (#1429)
the hairs are different for some reason please fix

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Half-elf hair colors are different from humen hair colors, despite
having the same names

## Why It's Good For The Game

Simply standardizing the colors. It's confusing when drywheat for a
half-elf is different from drywheat for a humen.

## Pre-Merge Checklist