Vampire Lord Enhancements (#1430)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Title, also increases vampire max blood by 50%. Playtesting shows they
typically run out of blood within a single day, and it's not ideal to
constantly need to kill people just to not starve to death.

This PR allows the vampire lord to innately summon a portal on their
location(900 second cooldown and duration). This portal when stepped on
allows any character, vampire or human to transport themselves to the
Vampire Castle(Specifically the Rift Gate). This allows the Lord and
their retinue to avoid the pesky issue of constantly dying in the
Murderwood on their way home. It also adds the unique RP opportunity for
people to besiege their castle via the portal, or be invited for lavish
feasts / parties.

Imagine a round where in which the Vampire Lord holds a masquerade ball,
the entrance to which is found in the upstairs room of the Tavern. The
guests are invited only to find themselves the night's dinner, with very
few being turned into vampires. The following day they descend on the
city, either through the portal(if it's still up) or via the murderwood
on a long trek toward the apocalypse.

## Why It's Good For The Game

Vampire is by the far the weakest antagonist, they often die from the
sun or because they run out of blood. It's incredibly arduous, and
annoying to play due to the extreme distances you must travel to reach
the castle. As a result they often just end up living in the sewers like
hobos, praying that the night will come so they can scrounge together
enough blood to barely survive until the next. They never, ever win. I
doubt this PR would make them strong enough to be a true threat, but at
least they can have fun now.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.