Gatemaster again because me stupid (#1433)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

* Returns the gatemaster job. A role dedicated to manning the keeps
gate.
* Adds three subclasses to the gatemaster. 
  1. Chainguard = ligth armor + chainwhip
  2. Brawler = Heavy armor + shield & mace
  3. Archer = Ligth armor + bow/crossbow
* Removes the gate helemt from city guards and turned it back into the
gate masters signature helmet
* Added a lot of additional manor guard spawns, to unclog them a bit
* Sligthly changes the gatehouse and to allow guards to patrol the roof
* removes the random bonus armor and weapons from the gatehouse
* Adds a chain whip. An iron whip that is around the lv of a golden
whip, with a bit more AP
* Adds a smithing recipy for the iron chain whip

## Why It's Good For The Game

The Gatemaster is a role that was lost as the codebase traveld through
multiple servers.
Now its back.

Its a dedicated role to guard the keeps gate. Allowing players that
actually want to do that, to main it.
And Man-at-arms, not to play rock paper sciccors for who has to be on
gate duty.
This should overall improve how serious the gate to the keep is guarded.
Making it more into a stable role, with set expectations for players
before they join a round.

And as the Gatemaster returns, they bring the chaiwhip with them for one
of their subclasses.
The Chainguard.
A heavy iron whip with a strg requirment of 9.
It is basicly a spider whip with 10+ AP, and double charge time on lash.
Been wanting to put a chain whip in, and it feels like it fits to the
gatemaster well.

The other sublcasses are 
An Archer. Wich likly spawns with a cross bow and unlikly with a bow.
And the Bruiser. A more heavily armored version that uses a mace and
shield.

I migth have given them a bit to much in some regard. So lemme know if
so.

Discord: NHC

Map changes
https://imgur.com/a/VJu2GxS

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->