Map votes now check to see if a given map is votable (#1525)
## About The Pull Request

Map votes currently do not take any part of the configuration into
account, aside from if the given map was disabled. This is bad because
certain maps need certain conditions to run. This PR implements a
`votable()` proc to remedy this issue.

## Why It's Good For The Game

Letting people vote for 80+ pop maps at 50 pop is not very good.