Rosewood shop adjustments + Adding the actual hotsprings (#1526)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

I have pretty much ripped up the tailor and swapped the tailors and
purchasable shops. IT IS NOT PRETTY. It is TEMPORARY I am testing how
moving the tailors location effects traffic flow. In addition I have
touched up the north roof of the keep and the town square with sprites
we have been waiting on. The tree is a placeholder

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->
Things being easy to find are good

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->