Soap, Buckets, and Cloth Changes (#1168)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

- ### Soap
- Bath Soap refactored into a subtype of normal soap. Also deleted the
accursed Nitemaiden file as a result.
- Soap can be dissolved into a container to get soap reagent. Soap
reagent makes cleaning more effective and can slip you. Can be used
alone to clean a floor, but can be ineffective.
  - New sprite for herbal soap
  - Crafting recipe added for herbal soap, requires mentha.
  - Soap spawns across the map changed to make herbal soap more rare.
- You can wash people's mouth out with soap if you grab them and aim for
their mouth with soap. You should absolutely do it over and over again
and see what happens
- Washing with good soap gives you a +1 fortune buff for 15 minutes.
Washing with either gives you a positive stress event.

- ### Cloth
- Now a pseudo reagent container with two new intents, soak and wring.
It can soak in reagents (including off the floor), and you can wring it
out. Holds 9 units.
- Uses the cleaner component now. Its base cleaning capability is pretty
poor. Soaking it in water and dirty water makes it more effective.
Soaking it with soapy water makes it even more so.
  - The wet system fucking blew so it's gone now.

- ### Buckets (But really all reagent containers)
- Splash intent actually does what you would expect it to! Performs a
chem splash on the selected tile
  - Increased the reach of splash by an extra tile and gave it sound.
  - Refactored cheese making
  - Milking a creature needs you to be on fill intent.

- ### Other Changes
  - Changed the clean values on a whole lot of cleanable effects

## Why It's Good For The Game

When I was 15, my brothers both came to me with a bar of soap with a
bite taken out of it. They asked me why I did this, and I told them I
didn't. They didn't believe me, and I never found out who took a bite
out of that soap. To this day, they think I took a bite out of that bar
of soap. Are they fucking with me? I don't know, and I don't think I
ever will.

This has nothing to do with the game.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: tonty <39193182+tontyGH@users.noreply.github.com>